feat: :building_construction: introduce Atomic Design
