+++ next.config.js
@@ -3,7 +3,7 @@ const nextConfig = {
   reactStrictMode: true,
   images: {
     loader: 'akamai',
-    path: ''
+    path: ' '
   },
 
   webpack(config) {
