+++ package.json
@@ -11,29 +11,29 @@
   },
   "dependencies": {
     "@kodingdotninja/use-tailwind-breakpoint": "^0.0.5",
-    "@svgr/webpack": "^6.2.1",
+    "@svgr/webpack": "^6.3.1",
     "animate.css": "^4.1.1",
-    "next": "12.0.8",
-    "next-themes": "^0.1.1",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-icons": "^4.3.1"
+    "next": "12.2.5",
+    "next-themes": "^0.2.0",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-icons": "^4.4.0"
   },
   "devDependencies": {
-    "@commitlint/cli": "^17.0.3",
-    "@commitlint/config-conventional": "^17.0.3",
-    "@types/node": "^17.0.8",
-    "@types/react": "^17.0.38",
-    "autoprefixer": "^9",
+    "@commitlint/cli": "^17.1.2",
+    "@commitlint/config-conventional": "^17.1.0",
+    "@types/node": "^18.7.14",
+    "@types/react": "^18.0.18",
+    "autoprefixer": "^10.4.8",
     "cz-conventional-changelog": "3.3.0",
-    "eslint": "8.6.0",
-    "eslint-config-next": "12.0.8",
+    "eslint": "8.23.0",
+    "eslint-config-next": "12.2.5",
     "husky": "^8.0.0",
     "lint-staged": "^13.0.3",
-    "postcss": "^7",
+    "postcss": "^8.4.16",
     "prettier": "^2.7.1",
     "tailwindcss": "npm:@tailwindcss/postcss7-compat",
-    "typescript": "^4.5.4"
+    "typescript": "^4.8.2"
   },
   "config": {
     "commitizen": {
