+++ src/components/index.ts
@@ -1,15 +1,3 @@
-export * from './Footer';
-export * from './Header';
-export * from './SocialButton';
-export * from './Hamburguer';
-export * from './MenuHamburguer';
-export * from './Card';
-export * from './DarkModeButton';
-export * from './Navbar';
-export * from './NavbarMobile';
-export * from './GoTo';
-export * from './GoToCard';
-export * from './JobsCard';
-export * from './UserButton';
-export * from './JobsSearchInput';
-export * from './JobsSearchForm';
+export * from './atoms';
+export * from './organisms';
+export * from './templates';
