+++ src/components/templates/index.ts
@@ -1,0 +1,2 @@
+export * from './Home';
+export * from './Jobs';
