+++ src/pages/jobs.ts
@@ -1,0 +1,3 @@
+import { Jobs } from '../components';
+
+export default Jobs;
